+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#1328](https://github.com/hyperf/hyperf/pull/1328) Added `ModelRewriteInheritanceVisitor` to rewrite the model inheritance for command `gen:model`.
 - [#1331](https://github.com/hyperf/hyperf/pull/1331) Added `Hyperf\LoadBalancer\LoadBalancerInterface::getNodes()`.
 - [#1335](https://github.com/hyperf/hyperf/pull/1335) Added event `AfterExecute` for `command`.
+- [#1361](https://github.com/hyperf/hyperf/pull/1361) Added config of `processors` for logger.
 
 ## Changed